+++ src/perpetual/clients.ts
@@ -2,7 +2,7 @@
  * Client models
  */
 
-import { X10BaseModel } from '../utils/model';
+import { X10BaseModel } from '../utils/model.js';
 
 /**
  * Client model